Small corrections to exercise mat_histogram

+++ mat_histogram.m
@@ -12,7 +12,7 @@ mat = im2uint8(mat);
 
 
 %% Initialize output vector
-max_num = intmax('uint8'); % this is safe because of the above conversion
+max_num = int32(intmax('uint8')); % this is safe because of the above conversion
 h = zeros(1, max_num + 1);
   
 %% Calculate histogram
